Create PYTHON CODE
This code snippet demonstrates a simple image encryption tool using pixel manipulation in Python. The tool allows users to encrypt and decrypt images by performing operations on each pixel. To use the tool, you need to have the Python Imaging Library (PIL) installed. You can install it using the following command:
Pip install pillow
The encrypt_image function takes the path of an image file as input and encrypts the image by performing an encryption operation on each pixel. The encrypted image is then saved as "encrypted_image.png" and the path is returned.The decrypt_image function takes the path of the encrypted image file as input and decrypts the image by performing a decryption operation on each pixel. The decrypted image is then saved as "decrypted_image.png" and the path is returned. To use the tool, you can provide the path of the image you want to encrypt as image_path. The encrypt_image function will return the path of the encrypted image. You can then pass this path to the decrypt_image function to get the path of the decrypted image. Please note that the encryption and decryption operations are not implemented in the code snippet. You need to define your own encryption and decryption algorithms based on your requirements.

Remember to handle exceptions and add error checking to ensure the code works correctly with different image formats and sizes.